+++ UI/assets/css/style.css
@@ -17,15 +17,24 @@ html, body{
     font-family:'Source Sans Pro', sans-serif;
     background: #fff;
     color: #000;
+   
+}
+body{
+    display: flex;
+    min-height: 100vh;
+    flex-direction: column;
+}
+.main{
+    flex: 1;
 }
 a{
-    color: #99d3c4;
+    color: #78d6be;
     text-decoration: none;
     cursor: pointer;
 }
 /*** COLOR & BACKGROUNDS ***/
-.brand-color{ color:#99d3c4; }
-.brand-bg{ background:#99d3c4; }
+.brand-color{ color:#78d6be; }
+.brand-bg{ background:#78d6be; }
 .blue-bg{ background: rgb(147, 228, 228, 0.7); }
 .text-dark{ color: #272727; }
 .text-black-2{ color:  rgba(71,71,71,.88); }
@@ -36,10 +45,12 @@ a{
 /*** BORDERS ***/
 .bd-blk{ border: 1px solid; }
 .bd-light{ border: 1px solid rgba(71,71,71,.88); }
+.bd-grey{ border: 1px solid #eee; }
 .bd-r-light{ border-right: 1px solid rgba(71,71,71,.88); }
 .bd-t-light{ border-top: 1px solid rgba(71,71,71,.88); }
 .rounded{ border-radius: 50%; }
-.shadow{ box-shadow: 1px 3px 5px 6px #e5e5e5; }
+.bd-none{ border-radius: 0; }
+.shadow{ box-shadow: 0px 0px 5px 4px #e5e5e5 }
 
 
 /*** WIDTH ***/
@@ -47,30 +58,36 @@ a{
 .w-100{ width: 100%; }
 .w-80p { width: 80%; }
 .w-50{ width: 50px; }
+.w-20{ width: 20px; }
 .h-100{ height: 100px !important; }
+.row{ display: flex; }
 .col-1{
-    display: inline-flex;
     width: 15%;
+    flex: 0 0 15%;
 }
 .col-2{ 
-    display: inline-flex; 
     width: 20%;
+    flex: 0 0 20%;
 }
 .col-3{
     display: inline-flex;
     width: 30%;
 }
+.col-4{
+    width: 33.3%;
+    flex: 0 0 33.3%;
+}
 .col-8{
-    display: inline-flex;
     width: 60%;
+    flex: 0 0 60%;
 }
 .col-9{
-    display: inline-flex;
     width: 75%;
+    flex: 0 0 75%;
 }
 .col-10{
-    display: inline-flex;
     width: 85%;
+    flex: 0 0 85%;
 }
 
 /*** PADDING ***/
@@ -89,7 +106,10 @@ a{
     padding-left: 3em;
     padding-right: 3em;
 }
-
+.py-2e{
+    padding-top: 2em;
+    padding-bottom: 2em;
+}
 
 /*** MARGIN ***/
 .mt-5{ margin-top: 20px; }
@@ -102,13 +122,14 @@ a{
 .mt-1e{ margin-top: 1em; }
 .mt-9e{ margin-top:10em; }
 .mx-auto{ margin: 0 auto;}
-
+.mr-1{ margin-right: 1em}
 .wrap{ flex-wrap: wrap; }
 
 
 /*** FONTS ***/
 .f-38{ font-size: 38px;}
 .f-24 { font-size: 24px; }
+.f-18{ font-size: 18px; }
 .f-12{ font-size: 12px; }
 .l--8{ line-height: 0.8; }
 .light{ font-weight: 300; }
@@ -133,18 +154,24 @@ a{
 .dis-inline{ display: inline-flex; }
 .dis-block{ display: block; }
 .list-unstyled{ list-style-type: none; }
+.justify-center{ justify-content: center; }
 
 
 /*** BUTTONS ***/
 .btn{
     border-radius: 0;
-    height: 37px;
-    width: 120px;
+    padding: 0.35em 1.5em;
     background: transparent
 }
+.btn:hover{
+    box-shadow: 3px 3px 0 #4d9f8a;
+    -webkit-transform: translate(-3px,-3px);
+    transform: translate(-3px,-3px);
+    outline: none;
+}
 .btn-outline-brand{
-    border: 2px solid #99d3c4 ;
-    color: #99d3c4 ;
+    border: 2px solid #78d6be ;
+    color: #78d6be ;
 }
 .btn-outline-white{
     border: 1px solid #fff;
@@ -153,16 +180,12 @@ a{
 .upvote-btn{
     padding: 0.4em 1em;
 }
-.upvote-btn i{
-    margin-right: 5px;
+.upvote-btn i, .downvote-btn i, .comment-btn i{
+    margin-right: 10px;
+    padding-top: 4px;
 }
 .upvote .upvote-count{
     margin-left: 5px;
-}
-.upvote:hover{
-    background: #99d3c4;
-    color: #fff !important;
-    text-transform: uppercase;
 }
 
 .downvote-btn{
@@ -172,8 +195,29 @@ a{
     color: #b06150;
     margin-right: 1em;
 }
-
+.upvote:hover{
+    background: #78d6be;
+    color: #fff;
+}
+.comment-btn:hover{
+    border: 1px solid #78d6be;
+}
 .downvote-btn:hover{
+    background: #b06150;
+    color: #fff;
+    box-shadow: 3px 3px 0 #4d261d;
+    -webkit-transform: translate(-3px,-3px);
+    transform: translate(-3px,-3px);
+    outline: none;
+    
+}
+.upvote-btn:hover i{
+    background: #78d6be;
+    color: #fff !important;
+    text-transform: uppercase;
+}
+
+.downvote-btn:hover i{
     background: #b06150;
     color: #fff;
 }
@@ -191,13 +235,13 @@ a{
 }
 .login-btn{
     padding: 10px 3em;
-    background: #99d3c4;
+    background: #78d6be;
     color: #fff;
     text-decoration: none;
 }
-.login-btn:hover, .signup-btn:hover{
+/* .login-btn:hover, .signup-btn:hover{
     background: #272727;
-}
+} */
 
 .header{
     background: #e5e5e5;
@@ -251,16 +295,17 @@ a{
     overflow: hidden;
 }
 .dropdown .dropbtn {
-    border: 1px solid #99d3c4;
+    border: 1px solid #78d6be;
     outline: none;
     color: #000;
     padding: 5px 15px;
     background-color: inherit;
     margin: 0;
+    font-size: 15px;
 }
   
 .dropdown:hover .dropbtn {
-    background-color: #99d3c4;
+    background-color: #78d6be;
 }
   
 .dropdown-content {
@@ -281,7 +326,7 @@ a{
 }
   
 .dropdown-content a:hover {
-    background-color: #99d3c4;
+    background-color: #78d6be;
     color: #fff;
 }
   
@@ -290,28 +335,28 @@ a{
 }
 
 .nav-list.active{
-    background: #99d3c4;
+    background: #78d6be;
 }
 .nav-item{
     text-decoration: none;
     color: #272727;
     font-weight: 300
 }
-.nav-list.active>.nav-item:hover{
+/* .nav-list.active>.nav-item:hover{
     color: #fff;
-}
+} */
 .nav-item:hover{
     text-transform: uppercase;
     font-size: 16px;
-    color: #99d3c4;
+    color: #fff;
     font-weight: 600;
 }
 .navbar-toggler{
     width: 50px;
     height: 50px;
-    color: #99d3c4;
+    color: #78d6be;
     background: transparent;
-    border: 1px solid #99d3c4;
+    border: 1px solid #78d6be;
 }
 .collapse{
     display: none;
@@ -326,7 +371,7 @@ a{
 
 .form-control{
     height: 37px;
-    width: 90%;
+    width: 100%;
     padding-left: 10px;
     padding-right: 10px;
     margin-top: 10px;
@@ -413,4 +458,11 @@ textarea[name='description']{
 }
 .comment-body p{
     margin: 0 10px 20px 10px;
+}
+.card-box{
+    padding: 4em 1.3em;
+}
+
+.card-box-container:hover{
+    box-shadow: 2px 3px 5px 2px #4d9f8a;
 }
