+++ UI/index.html
@@ -50,7 +50,7 @@
         <!-- ./ header -->
 
         <!-- main -->
-        <main>
+        <main class="main">
             <div class="container">
                 <div class="wrapper mt-9e text-center">
                     <h3 class="f-38 l--8 light">
@@ -58,7 +58,11 @@
                         
                     </h3>
                    <h2 class="mt-2e f-24 light mb-2e l--3"> <span class="brand-color">​Questioner​​</span>  helps the meetup organizers prioritize questions to be answered</h2>
-                   <a href="login.html" class="btn login-btn"> <strong>Login</strong> </a>
+                   <div>
+                        <a href="register.html" class="btn login-btn mr-1"> <strong>Sign Up</strong> </a>
+                        <a href="login.html" class="btn login-btn"> <strong>Login</strong> </a>
+                   </div>
+                   
                 </div>
                 
             </div>
