+++ UI/profile.html
@@ -49,14 +49,14 @@
         <!-- ./ header -->
 
         <!-- main -->
-        <main>
+        <main class="main">
             <div class="container">
-                <div class="wrapper mt-9e text-center dis-flex space-even">
+                <div class="row mt-9e text-center space-even">
                     <div class="col-3">
-                        <div class="w-100">
+                        <div class="">
                             <div class="profile-card">
-                                <div class="brand-bg img-card pos-rel">
-                                    <img src="assets/img/user-profile-2.jpg" class="user-profile-img pos-abs r-38p" alt="">
+                                <div class="brand-bg img-card pos-rel dis-flex justify-center">
+                                    <img src="assets/img/user-profile-2.jpg" class="user-profile-img pos-abs" alt="">
                                 </div>
                                 <div class="card-body mt-5e px-2">
                                     <div class="user-details">
@@ -78,37 +78,44 @@
                                         </a>
                                     </div>
                                 </div>
-                                <div class="card-footer bd-t-light mt-5 px-2 dis-flex space-even wrap">
-                                    <div class="bd-r-light pt-1e pb-1e pr-2e text-center">
-                                        <h3>
-                                            200
-                                        </h3>
-                                        <p class="text-black-2 text-uppercase"> Asked</p>
-                                    </div>
-
-                                    <div class="bd-r-light pt-1e pb-1e pr-2e">
-                                        <h3>
-                                            200
-                                        </h3>
-                                        <p class="text-black-2 text-uppercase"> comments</p>
-                                    </div>
-                                    <div class="pt-1e pb-1e px-2">
-                                        <h3>
-                                            200
-                                        </h3>
-                                        <p class="text-black-2 text-uppercase"> upvotes</p>
+                                <div class="card-footer bd-t-light mt-5 px-2">
+                                    <div class="row wrap">
+                                        <div class="col-4">
+                                            <div class="bd-r-light pt-1e pb-1e pr-2e text-center">
+                                                <h3>
+                                                    200
+                                                </h3>
+                                                <p class="text-black-2 text-uppercase"> Asked</p>
+                                            </div>
+                                        </div>
+                                        <div class="col-4">
+                                            <div class="bd-r-light pt-1e pb-1e pr-2e">
+                                                <h3>
+                                                    200
+                                                </h3>
+                                                <p class="text-black-2 text-uppercase"> comments</p>
+                                            </div>
+                                        </div>
+                                        <div class="col-4">
+                                            <div class="pt-1e pb-1e px-2">
+                                                <h3>
+                                                    200
+                                                </h3>
+                                                <p class="text-black-2 text-uppercase"> upvotes</p>
+                                            </div>
+                                        </div>
                                     </div>
                                 </div>
                             </div>
                         </div>
                     </div>
                     <div class="col-8">
-                        <div class="w-100 px-2 text-left pt-1e shadow pb-1e activity-container">
-                            <h3>Activity</h3>
+                        <div class="px-2 text-left pt-1e shadow pb-1e activity-container">
+                            <h3>Activities</h3>
                             <div class="activities dis-flex space-around mt-3e">
                                 <div class="col-1">
                                     <div class="rounded mx-auto">
-                                        <img src="assets/img/meetup.png" alt="" class="rounded w-50 bd-light">
+                                        <img src="assets/img/m-logo-3.jpg" alt="" class="rounded w-50 bd-light">
                                     </div> 
                                 </div>
                                 <div class="col-10">
@@ -138,7 +145,7 @@
                             <div class="activities dis-flex space-around mt-3e">
                                 <div class="col-1">
                                     <div class="rounded mx-auto">
-                                        <img src="assets/img/meetup.png" alt="" class="rounded w-50 bd-light">
+                                        <img src="assets/img/m-logo-1.png" alt="" class="rounded w-50 bd-light">
                                     </div> 
                                 </div>
                                 <div class="col-10">
@@ -168,7 +175,7 @@
                             <div class="activities dis-flex space-around mt-3e">
                                 <div class="col-1">
                                     <div class="rounded mx-auto">
-                                        <img src="assets/img/meetup.png" alt="" class="rounded w-50 bd-light">
+                                        <img src="assets/img/m-logo-2.jpg" alt="" class="rounded w-50 bd-light">
                                     </div> 
                                 </div>
                                 <div class="col-10">
@@ -198,7 +205,7 @@
                             <div class="activities dis-flex space-around mt-3e">
                                 <div class="col-1">
                                     <div class="rounded mx-auto">
-                                        <img src="assets/img/meetup.png" alt="" class="rounded w-50 bd-light">
+                                        <img src="assets/img/m-logo-3.jpg" alt="" class="rounded w-50 bd-light">
                                     </div> 
                                 </div>
                                 <div class="col-10">
