+++ UI/register.html
@@ -96,7 +96,7 @@
                         </div>
                         <div class="form-group mt-5e text-center">
                             <button  type="submit" class="btn btn-outline-white signup-btn" id="signup_btn"><strong>Sign Up</strong> </button>
-                            <p class="text-white mt-2e">Already have an account? Click <a href="register.html" class="text-dark">here</a> to log in</p>
+                            <p class="text-white mt-2e">Already have an account? Click <a href="login.html" class="text-dark">here</a> to log in</p>
                         </div>
                     </form>
                 </div>
