+++ app.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 
-import Routes from './server/src/routes/index';
+import Routes from './server/src/v1/routes/index';
 
 
 const app = express();
