+++ package.json
@@ -4,9 +4,9 @@
   "description": "A Full stack application that Crowd-source questions for a meetup. ​Questioner​​ helps the meetup organizer prioritize  questions to be answered.",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon --exec babel-node ./server/src/lib/server.js",
+    "start": "nodemon --exec babel-node ./server/src/v1/lib/server.js",
     "lint": "eslint",
-    "test": "mocha --require @babel/register ./server/src/tests/questions.spec.js",
+    "test": "mocha --require @babel/register ./server/src/v1/tests/meetups.spec.js",
     "coveralls": "nyc npm test && nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
