+++ server/src/v1/lib/server.js
@@ -1,0 +1,10 @@
+import http from 'http';
+
+import app from '../../../../app';
+
+const port = process.env.PORT || 8000;
+const server = http.createServer(app);
+
+server.listen(port, () => {
+    console.log(`server is running at port ${port}`);
+});
