+++ server/src/v1/models/meetupRecord.js
@@ -1,0 +1,64 @@
+const meetupRecord = [
+    {
+        id: 1,
+        meetupId: 1011,
+        createdOn: new Date(),
+        location: 'lagos',
+        imageUrl: 'placeholder/image.com',
+        topic: 'Gryffindor meetup',
+        happeningOn: new Date(),
+        tags: ['house gryffindor', 'dumbledore army']
+    },
+    {
+        id: 2,
+        meetupId: 1011,
+        createdOn: new Date(),
+        location: 'lagos',
+        imageUrl: 'placeholder/image.com',
+        topic: 'Gryffindor meetup',
+        happeningOn: new Date(),
+        tags: ['house gryffindor', 'dumbledore army']
+    },
+    {
+        id: 3,
+        meetupId: 1011,
+        createdOn: new Date(),
+        location: 'lagos',
+        imageUrl: 'placeholder/image.com',
+        topic: 'Gryffindor meetup',
+        happeningOn: new Date(),
+        tags: ['house gryffindor', 'dumbledore army']
+    },
+    {
+        id: 4,
+        meetupId: 1011,
+        createdOn: new Date(),
+        location: 'lagos',
+        imageUrl: 'placeholder/image.com',
+        topic: 'Gryffindor meetup',
+        happeningOn: new Date(),
+        tags: ['house gryffindor', 'dumbledore army']
+    },
+    {
+        id: 5,
+        meetupId: 1011,
+        createdOn: new Date(),
+        location: 'lagos',
+        imageUrl: 'placeholder/image.com',
+        topic: 'Gryffindor meetup',
+        happeningOn: new Date(),
+        tags: ['house gryffindor', 'dumbledore army']
+    },
+    {
+        id: 6,
+        meetupId: 1011,
+        createdOn: new Date(),
+        location: 'lagos',
+        imageUrl: 'placeholder/image.com',
+        topic: 'Gryffindor meetup',
+        happeningOn: new Date(),
+        tags: ['house gryffindor', 'dumbledore army']
+    }
+];
+
+export default meetupRecord;
