+++ server/src/v1/models/questionsRecord.js
@@ -1,0 +1,21 @@
+const questionsRecord = [
+    {
+        id: 1,
+        meetupId: 10011,
+        meetupTitle: 'Gryffindor meets',
+        title: 'Enquiries',
+        body: 'When is the next meeting',
+        votes: 1,
+    },
+    {
+        id: 2,
+        meetupId: 10011,
+        meetupTitle: 'Gryffindor meets',
+        title: 'Enquiries',
+        body: 'When is the next meeting',
+        votes: 1
+    },
+];
+
+
+export default questionsRecord;
